+++ runtime/core/xnix_runtime_core.c
@@ -1993,6 +1993,7 @@ xnix_runtime_compatibility_action_queue(
 #include "xnix_runtime_core_tray_status.inc"
 #include "xnix_runtime_core_krunner_query.inc"
 #include "xnix_runtime_core_portal_request.inc"
+#include "xnix_runtime_core_compatibility_repair_plan.inc"
 #include "xnix_runtime_core_ai_diagnostic_input.inc"
 #include "xnix_runtime_core_ai_diagnostic_recommendation.inc"
 #include "xnix_runtime_core_ai_repair_approval_gate.inc"