+++ runtime/core/xnix_runtime_core_cli.c
@@ -2818,6 +2818,8 @@ print_probe(void)
   fputs("\"compatibility_center_summary_count\":1,", stdout);
   fputs("\"compatibility_run_plan_owner\":\"c\",", stdout);
   fputs("\"compatibility_run_plan_count\":1,", stdout);
+  fputs("\"compatibility_repair_plan_owner\":\"c\",", stdout);
+  fputs("\"compatibility_repair_issue_count\":4,", stdout);
   fputs("\"desktop_activation_manifest_owner\":\"c\",", stdout);
   fputs("\"desktop_activation_manifest_count\":1,", stdout);
   fputs("\"kde_integration_status_owner\":\"c\",", stdout);
@@ -2908,6 +2910,7 @@ print_write_gate(const char *method_name)
 #include "xnix_runtime_core_cli_tray_status.inc"
 #include "xnix_runtime_core_cli_krunner_query.inc"
 #include "xnix_runtime_core_cli_portal_request.inc"
+#include "xnix_runtime_core_cli_compatibility_repair_plan.inc"
 #include "xnix_runtime_core_cli_ai_diagnostic_input.inc"
 #include "xnix_runtime_core_cli_ai_diagnostic_recommendation.inc"
 #include "xnix_runtime_core_cli_ai_repair_approval_gate.inc"